Highcharts - public api
