+++ libs/sdk-ui/src/base/helpers/common.ts
@@ -7,8 +7,7 @@ import { Observable } from "rxjs/Rx";
 import { numberFormat } from "@gooddata/numberjs";
 
 import { VisualizationTypes } from "../constants/visualizationTypes";
-import { IAxis, ISeriesItem, IChartOptions } from "../../interfaces/Config";
-import { IHighchartsAxisExtend } from "../../highcharts/HighchartsExtend";
+import { ISeriesItem, IChartOptions } from "../../interfaces/Config";
 
 // lodash/fp does not provide typings
 // https://stackoverflow.com/questions/38020019/where-can-i-find-typescript-typings-for-lodash-fp
@@ -68,7 +67,6 @@ export const isFunnelChart = isEqual(VisualizationTypes.FUNNEL);
 export const isHeatmap = isEqual(VisualizationTypes.HEATMAP);
 export const isChartSupported = (type: string) => includes(VisualizationTypes, type);
 export const isOneOfTypes = (type: string, types: string[]) => includes(types, type);
-export const isPrimaryYAxis = (yAxis: IAxis | IHighchartsAxisExtend) => !yAxis.opposite;
 export const stringifyChartTypes = () =>
     Object.keys(VisualizationTypes)
         .reduce((acc, type) => {
