+++ libs/sdk-ui/src/base/helpers/executionResultHelper.ts
@@ -1,6 +1,6 @@
 // (C) 2007-2018 GoodData Corporation
 import * as invariant from "invariant";
-import { IUnwrappedAttributeHeadersWithItems } from "../../highcharts/chart/types";
+import { IUnwrappedAttributeHeadersWithItems } from "./types";
 import { IMappingHeader } from "../../interfaces/MappingHeader";
 import { getMappingHeaderLocalIdentifier } from "./mappingHeader";
 import { IAttributeHeader, IMeasureGroupHeader, IResultDimension } from "@gooddata/sdk-backend-spi";
