+++ libs/sdk-ui/src/base/helpers/types.ts
