+++ libs/sdk-ui/src/highcharts/chart/highcharts/highchartsEntryPoint.ts
@@ -3,4 +3,11 @@
 // Import this reexported variable in other files instead of direct import from highcharts
 import * as Highcharts from "highcharts";
 
+export type HTMLDOMElement = Highcharts.HTMLDOMElement;
+export type SVGDOMElement = Highcharts.SVGDOMElement;
+export type SVGAttributes = Highcharts.SVGAttributes;
+export type StackItemObject = Highcharts.StackItemObject;
+export type DataLabelsOptionsObject = Highcharts.DataLabelsOptionsObject;
+export type ColorAxisOptions = Highcharts.ColorAxisOptions;
+
 export default Highcharts;
