+++ libs/sdk-ui/src/highcharts/index.ts
@@ -1,0 +1,20 @@
+import Chart from "./chart/Chart";
+import ChartTransformation from "./chart/ChartTransformation";
+
+export { Visualization } from "./Visualization";
+
+export { Chart, ChartTransformation };
+
+export {
+    HTMLDOMElement,
+    SVGDOMElement,
+    SVGAttributes,
+    StackItemObject,
+    DataLabelsOptionsObject,
+    ColorAxisOptions,
+} from "./chart/highcharts/highchartsEntryPoint";
+
+export { AfmPropTypesShape, ResultSpecPropTypesShape, FiltersPropTypesShape } from "./proptypes/execution";
+export { FLUID_LEGEND_THRESHOLD } from "./chart/HighChartsRenderer";
+export { COMBO_SUPPORTED_CHARTS } from "./chart/chartOptions/comboChartOptions";
+export { createDrillIntersectionElement } from "./utils/drilldownEventing";
