+++ libs/sdk-ui/src/highcharts/utils/isPrimaryYAxis.ts
@@ -1,0 +1,4 @@
+import { IAxis } from "../../interfaces/Config";
+import { IHighchartsAxisExtend } from "../HighchartsExtend";
+
+export const isPrimaryYAxis = (yAxis: IAxis | IHighchartsAxisExtend) => !yAxis.opposite;
