+++ libs/sdk-ui/src/index.ts
@@ -44,8 +44,7 @@ import { Executor } from "./execution/Executor";
 import * as ChartConfiguration from "./interfaces/Config";
 // tslint:disable-next-line:no-duplicate-imports
 import { ILegendConfig, IChartConfig, IColorPalette, IColorPaletteItem } from "./interfaces/Config";
-import Chart from "./highcharts/chart/Chart";
-import ChartTransformation from "./highcharts/chart/ChartTransformation";
+import { Chart, ChartTransformation } from "./highcharts";
 import { RuntimeError } from "./base/errors/RuntimeError";
 import { IMeasureTitleProps, IArithmeticMeasureTitleProps } from "./interfaces/MeasureTitle";
 import { OverTimeComparisonType, OverTimeComparisonTypes } from "./interfaces/OverTimeComparison";
