+++ libs/sdk-ui/src/proptypes/Execute.ts
@@ -1,6 +1,6 @@
 // (C) 2007-2018 GoodData Corporation
 import * as PropTypes from "prop-types";
-import { AfmPropTypesShape, ResultSpecPropTypesShape } from "../highcharts/proptypes/execution";
+import { AfmPropTypesShape, ResultSpecPropTypesShape } from "../highcharts";
 import { EventsPropTypes } from "./Events";
 
 import { Requireable } from "prop-types"; // tslint:disable-line:no-duplicate-imports
