+++ libs/sdk-ui/src/proptypes/Filters.ts
@@ -1,4 +1,4 @@
 // (C) 2007-2018 GoodData Corporation
-import { FiltersPropTypesShape } from "../highcharts/proptypes/execution";
+import { FiltersPropTypesShape } from "../highcharts";
 
 export const FiltersPropType: any = FiltersPropTypesShape;
