+++ libs/sdk-ui/stories/core_components/ColumnChart.tsx
@@ -25,7 +25,7 @@ import { ScreenshotReadyWrapper, createHighChartResolver } from "../utils/Screen
 
 import * as HeaderPredicateFactory from "../../src/base/factory/HeaderPredicateFactory";
 import { wrap } from "../utils/wrap";
-import { Visualization } from "../../src/highcharts/Visualization";
+import { Visualization } from "../../src/highcharts";
 import { dualChartWithComputedAttribute } from "../test_data/fixtures";
 
 const wrapperStyle = { width: 800, height: 400 };
