+++ libs/sdk-ui/stories/internal/ChartProperties.tsx
@@ -5,7 +5,7 @@ import { screenshotWrap } from "@gooddata/test-storybook";
 import identity = require("lodash/identity");
 import cloneDeep = require("lodash/cloneDeep");
 
-import ChartTransformation from "../../src/highcharts/chart/ChartTransformation";
+import { ChartTransformation } from "../../src/highcharts";
 import { VIEW_BY_DIMENSION_INDEX } from "../../src/base/constants/dimensions";
 import { BASE_DUAL_AXIS_CHARTS } from "../data/dualAxis";
