+++ libs/sdk-ui/stories/internal/ChartTransformation.tsx
@@ -4,8 +4,7 @@ import { storiesOf } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 import { screenshotWrap } from "@gooddata/test-storybook";
 
-import ChartTransformation from "../../src/highcharts/chart/ChartTransformation";
-import { FLUID_LEGEND_THRESHOLD } from "../../src/highcharts/chart/HighChartsRenderer";
+import { ChartTransformation, FLUID_LEGEND_THRESHOLD } from "../../src/highcharts";
 import { immutableSet } from "../../src/base/helpers/common";
 import { STACK_BY_DIMENSION_INDEX, VIEW_BY_DIMENSION_INDEX } from "../../src/base/constants/dimensions";
