+++ libs/sdk-ui/stories/internal/Drilldown.tsx
@@ -6,7 +6,7 @@ import { action, decorateAction } from "@storybook/addon-actions";
 import { screenshotWrap } from "@gooddata/test-storybook";
 import { Execution } from "@gooddata/gd-bear-model";
 
-import { Visualization } from "../../src/highcharts/Visualization";
+import { Visualization } from "../../src/highcharts";
 import * as headerPredicateFactory from "../../src/base/factory/HeaderPredicateFactory";
 import { wrap } from "../utils/wrap";
 import * as fixtures from "../test_data/fixtures";
