+++ libs/sdk-ui/stories/internal/DualAxesChartConfig.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 import { storiesOf } from "@storybook/react";
 import { screenshotWrap } from "@gooddata/test-storybook";
 import * as dataSet from "../test_data/chart_with_2_metrics_and_view_by_attribute";
-import { Visualization } from "../../src/highcharts/Visualization";
+import { Visualization } from "../../src/highcharts";
 import { wrap } from "../utils/wrap";
 import "../../styles/scss/charts.scss";
 import { ChartType, IChartConfig, VisualizationTypes } from "../../src";
