+++ libs/sdk-ui/stories/utils/CustomLegend.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 import { screenshotWrap } from "@gooddata/test-storybook";
 import identity = require("lodash/identity");
 
-import ChartTransformation from "../../src/highcharts/chart/ChartTransformation";
+import { ChartTransformation } from "../../src/highcharts";
 import { barChartWith3MetricsAndViewByAttribute } from "../test_data/fixtures";
 import { wrap } from "./wrap";
